Make the api.auth.Auth a Protocol
